% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{CLR transformation of community matrix, with multiplicative zero replacement}
\usage{
clr(mat, delta = 1)
}
\arguments{
\item{mat}{A community matrix with features as rows}

\item{delta}{The pseudocount to exchange zeroes with. Zero-correction is multiplicative such that the log-ratios between any entirely non-zero features will not be affected by the pseudocount.}
}
\value{
CLR transformed community matrix
}
\description{
This function performs a CLR transformation on the input.
Prior to transformation it does a multiplicative zero replacement, which adds a pseudocount, but corrects the non-zero abundances such that log-ratios between non-zero elements are unchanged after correction.
}
\keyword{clr}
\keyword{pseudocount}
