% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comdist.par.R
\name{comdist.par}
\alias{comdist.par}
\title{Inter-community mean pairwise distance}
\usage{
comdist.par(comm, dis, abundance.weighted = FALSE, cores = 1,
  progress = TRUE)
}
\arguments{
\item{comm}{Community data matrix with samples as rows}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{abundance.weighted}{Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)}

\item{cores}{How many cores should be used for parallel computing}

\item{progress}{Show a progress bar}
}
\value{
A distance matrix of MPD values
}
\description{
Parallel calculation of MPD (mean pairwise distance) separating taxa in two communities, a measure of phylogenetic beta diversity
}
\keyword{MPD}
\keyword{bMPD}
\keyword{bNRI}
\keyword{betaMPD}
\keyword{betaNRI}
