% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comdistnt.par.R
\name{comdistnt.par}
\alias{comdistnt.par}
\title{Inter-community mean nearest taxon distance}
\usage{
comdistnt.par(
  comm,
  dis,
  abundance.weighted = FALSE,
  exclude.conspecifics = FALSE,
  cores = 1,
  progress = TRUE
)
}
\arguments{
\item{comm}{Community data matrix with samples as rows}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{abundance.weighted}{Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)}

\item{exclude.conspecifics}{Should conspecific taxa in different communities be exclude from MNTD calculations? (default = FALSE)}

\item{cores}{How many cores should be used for parallel computing'}

\item{progress}{Show a progress bar}
}
\value{
A distance matrix of MNTD values
}
\description{
Parallel calculation of MNTD (mean nearest taxon distance) separating taxa in two communities, a measure of phylogenetic beta diversity
}
\keyword{MNTD}
\keyword{bMNTD}
\keyword{bNTI}
\keyword{betaMNTD}
\keyword{betaNTI}
