% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community_rrna.R
\name{community_rrna}
\alias{community_rrna}
\title{Calculate the community/sample-wise mean 16S rRNA copy numbers}
\usage{
community_rrna(x, copy.database = "v13.5", weighted = TRUE)
}
\arguments{
\item{x}{A phyloseq object OR an OTU-table with taxa as rows and OTU names as rownames. OTUs should be picked against the Greengenes v13.5 database, unless a another copy number database is provided.}

\item{copy.database}{What Greengenes database version was used to find OTUs. Atm only "v13.5" is available. Alternatively, A dataframe with two variables: "ID" is the OTU id matched by rownames in x and "Copy" is the copy number.}

\item{weighted}{Logical. Should the average copy number be weighted by the relative abundances of the OTUs? If not, it is adviced to rarefy the otu-table first if there are large differences in sample reads.}
}
\value{
A dataframe with an average copy number for each sample
}
\description{
Calculate the average 16S rRNA copy number for each sample in an OTU table
}
\keyword{rrna}
