% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_13_5_16S.R
\docType{data}
\name{gg_13_5_16S}
\alias{gg_13_5_16S}
\title{16S copy numbers for GreenGenes 13_5 database OTUs}
\description{
Precalculated for the PICRUSt pipeline (https://picrust.github.io/picrust/picrust_precalculated_files.html)
}
\keyword{data}
