% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neutral.rand.R
\name{neutral.rand}
\alias{neutral.rand}
\title{Fit Sloan et al. (2006) Neutral Model several times}
\usage{
neutral.rand(
  data,
  n = NULL,
  s = NULL,
  rRNA = NULL,
  rn = NULL,
  cores = 1,
  naming = NULL
)
}
\arguments{
\item{data}{A phyloseq object}

\item{n}{Integer. Number of times to repeat analysis}

\item{s}{Integer. Number of random samples to for each repetition.}

\item{rRNA}{What Greengenes database version was used to find OTUs. Atm only "v13.5" is available. Alternatively, A dataframe with two variables: "ID" is the OTU id matched by names in x and "Copy" is the copy number.}

\item{rn}{Integer. Number of reads to sample for rarefaction}

\item{cores}{Integer. Number of cores to use for parallel computing.}

\item{naming}{Optional. A list for naming the output e.g. list(Time="1 Week",Type="Gut")}
}
\value{
A list of length two; first element contains fit statistics, the second element contains predictions.
}
\description{
Fit neutral model developed by Sloan et al. (2006, Environ Microbiol 8(4):732-740) and implemented by Burns et al. (2015, ISME J 10(3):655-664) several times on ramdomly picked samples and with 16S rRNA gene copy number corrected rarefaction.
Outputs three different pseudo R2 values for model fit:
- McFadden's R2: 1 - (logLik(model) / logLik(null model))
- Cox & Snell's R2: 1 - exp((2/n) * (logLik(null model) - logLik(model)))
- Nagelkerke's R2: Cox & Snell's R2 / (1 - exp((2/n) * logLik(null model)))
}
\keyword{neutral}
