% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_euler.R
\name{phylo_euler}
\alias{phylo_euler}
\title{Make Euler (Venn) diagram of shared taxa (ASVs, OTUs) across sample groups}
\usage{
phylo_euler(ps, group, weight = FALSE, type = "percent", relative = TRUE, ...)
}
\arguments{
\item{ps}{A phyloseq object}

\item{group}{The grouping factor. Should match variable in sample_data(ps)}

\item{weight}{If TRUE, the overlaps are weighted by abundance}

\item{type}{"percent" or "count"}

\item{relative}{Should abundances be made relative}

\item{...}{Additional arguments}
}
\value{
An euler plot
}
\description{
Make Euler (Venn) diagram of shared taxa (ASVs, OTUs) across sample groups from a phyloseq object. Overlap can be weighted by relative abundance
}
\keyword{diagram}
\keyword{euler}
\keyword{venn}
