% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/proportionality.R
\name{proportionality}
\alias{proportionality}
\title{Calculate proportionality between all pairs of OTUs, for network analysis}
\usage{
proportionality(x, delta = 1)
}
\arguments{
\item{x}{An otu-table with OTUs as rows OR a phyloseq object}

\item{delta}{The pseudocount to exchange zeroes with. Zero-correction is multiplicative such that the proportionality between any entirely non-zero OTUs will not be affected by the pseudocount.}
}
\value{
The network association matrix with proportionality values
}
\description{
Calculate proportionality as proposed by Lovell et al. 2016 Proportionality: a valid alternative to correlation for relative data
}
\keyword{association}
\keyword{network}

