% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_euler.R
\name{ps_euler}
\alias{ps_euler}
\title{Make Euler diagram of shared taxa (ASVs, OTUs) across sample groups}
\usage{
ps_euler(
  ps,
  group,
  fraction = 0,
  shape = "circle",
  weight = FALSE,
  relative = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{ps}{A phyloseq object}

\item{group}{The grouping factor. Should match variable in sample_data(ps)}

\item{fraction}{The fraction (0 to 1) of samples in a group in which the taxa should be present to be included in the count.}

\item{shape}{Shape of the plot "circle" or "ellipse"?}

\item{weight}{If TRUE, the overlaps are weighted by abundance}

\item{relative}{Should abundances be made relative}

\item{plot}{If TRUE return a plot, if FALSE return a list with shared and unique taxa}

\item{...}{Additional arguments}
}
\value{
An euler plot
}
\description{
Make Euler diagram of shared taxa (ASVs, OTUs) across sample groups from a phyloseq object. Overlap can be weighted by relative abundance
}
\details{
Any further arguments to this function are passed to the plot.euler function from the eulerr package. This can be used to change colors, fonts, and other graphical parameters.
For example: ps_euler(phy, "Time", quantities = list(type=c("percent","counts"), font = 2), labels = list(cex = 2), col = "red", fill = c("red","blue","green"))
}
\keyword{diagram}
\keyword{euler}
