% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_pheatmap.R
\name{ps_pheatmap}
\alias{ps_pheatmap}
\title{Make pheatmap directly from a phyloseq object}
\usage{
ps_pheatmap(
  ps,
  annot_samp = NULL,
  annot_taxa = NULL,
  relative = TRUE,
  log10 = TRUE,
  tax_agg = NULL,
  order_taxa = TRUE,
  min_samples = 1,
  min_reads = 1,
  min_abundance = 0,
  label_rank = NULL,
  color = c("black", viridis::viridis(10)),
  ...
)
}
\arguments{
\item{ps}{A phyloseq object}

\item{annot_samp}{Sample variables to annotate with}

\item{annot_taxa}{Taxa variables to annotate with}

\item{relative}{If TRUE, abundances total sum scaled. Default TRUE}

\item{log10}{If TRUE, log10 transform abundances. Default TRUE}

\item{tax_agg}{Taxa rank to agglomorate. Default NULL}

\item{order_taxa}{If TRUE, taxa are ordered from most to least abundant. Default TRUE}

\item{min_samples}{Minimum number of samples the features should be present in. Default 1}

\item{min_reads}{Minimum number of total reads the features should have. Default 1}

\item{min_abundance}{Minimum mean relative abundance features should have. Default 0}

\item{label_rank}{Taxa rank to label the taxa. If NULL will label by taxa_names(ps). Default NULL}

\item{color}{Color palette. Default viridis with absent}

\item{...}{Additional arguments to pheatmap function}
}
\value{
A pheatmap
}
\description{
Make pheatmap directly from a phyloseq object, including agglomoration and filtering. 
Default color map is viridis with absent taxa as black.
}
