% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_prune.R
\name{ps_prune}
\alias{ps_prune}
\title{Prune taxa from phyloseq object by their prevalence or abundance}
\usage{
ps_prune(data, min.samples = 0, min.reads = 0, min.abundance = 0)
}
\arguments{
\item{data}{\code{phyloseq} object.}

\item{min.samples}{Minimum number of samples the features should be present in. Default 0}

\item{min.reads}{Minimum number of total reads the features should have. Default 0}

\item{min.abundance}{Minimum mean relative abundance features should have. Default 0}
}
\value{
Similar to input, but with features not reaching the criteria given grouped as "Others"
}
\description{
Prune taxa from phyloseq object by their prevalence or abundance
}
