% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_refactor.R
\name{ps_refactor}
\alias{ps_refactor}
\title{Relevel the Sample variable in a psmelted phyloseq object}
\usage{
ps_refactor(psmelted, ...)
}
\arguments{
\item{psmelted}{A phyloseq object melted into a data.frame with psmelt}

\item{...}{Arguments passed to hclust}
}
\description{
Relevel the Sample variable in a psmelted phyloseq object, 
such that similar samples are plotted together with ggplot barcharts
}
