% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_tax_clean.R
\name{ps_tax_clean}
\alias{ps_tax_clean}
\title{Clean tax_table such that NAs are replaced with names of the most specific known taxonomy prefixed with the rank.}
\usage{
ps_tax_clean(data)
}
\arguments{
\item{data}{\code{phyloseq} or \code{tax_table} object.}
}
\value{
Similar to input, but with NAs replaced in the tax_table
}
\description{
Example: An OTU is annotated as Proteobacteria in the Phylum rank, but NAs in the more specific ranks.
All annotations below Phylum for this OTU will be replaced with Phylum_Proteobacteria.
}
