% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_venn.R
\name{ps_venn}
\alias{ps_venn}
\title{Make Venn diagram of shared taxa (ASVs, OTUs) across sample groups}
\usage{
ps_venn(ps, group, weight = FALSE, type = "percent", relative = TRUE, ...)
}
\arguments{
\item{ps}{A phyloseq object}

\item{group}{The grouping factor. Should match variable in sample_data(ps)}

\item{weight}{If TRUE, the overlaps are weighted by abundance}

\item{type}{"percent" or "counts"}

\item{relative}{Should abundances be made relative}

\item{...}{Additional arguments}
}
\value{
An venn plot
}
\description{
Make Venn diagram of shared taxa (ASVs, OTUs) across sample groups from a phyloseq object. Overlap can be weighted by relative abundance
}
\keyword{diagram}
\keyword{venn}
