% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rarefy_rrna.R
\name{rarefy_rrna}
\alias{rarefy_rrna}
\alias{rarefy_rrna.matrix}
\alias{rarefy_rrna.phyloseq}
\title{Rarefy and normalize based on 16S rRNA copy numbers}
\usage{
rarefy_rrna(x, reads, copy.database = "v13.5", seed = NULL)

rarefy_rrna.matrix(x, reads, rRNA, seed = NULL)

rarefy_rrna.phyloseq(x, reads, rRNA, seed = NULL)
}
\arguments{
\item{x}{An otu-table with OTUs as columns OR a phyloseq object}

\item{reads}{Number of reads to sample.}

\item{copy.database}{What Greengenes database version was used to find OTUs. Atm only "v13.5" is available. Alternatively, A dataframe with two variables: "ID" is the OTU id matched by names in x and "Copy" is the copy number.}

\item{seed}{Random seed for sampling.}
}
\value{
A rarefied otu-table
}
\description{
Rarefy an otu-table with the probability of the inverse 16S rRNA copy numbers: The result is a standardized AND normalized otu-table.
}
\keyword{normalize}
\keyword{rarefy}
\keyword{rrna}

