% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurve.R
\name{rcurve}
\alias{rcurve}
\title{Rarefaction curve on phyloseq object
Theoretical richness with vegan's rarefy function}
\usage{
rcurve(physeq, subsamp = 10^c(1:5), trim = TRUE,
  add_sample_data = TRUE)
}
\arguments{
\item{physeq}{phyloseq object}

\item{subsamp}{Vector of number of reads to subsample}

\item{trim}{Remove richness estimations from subsamples larger than the library size}

\item{add_sample_data}{Add sample data to data.frame}
}
\description{
Rarefaction curve on phyloseq object
Theoretical richness with vegan's rarefy function
}
