% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.comdist.R
\name{ses.comdist}
\alias{ses.comdist}
\title{Standardized effect size of inter-community MPD (betaMPD, betaNRI)}
\usage{
ses.comdist(samp, dis, null.model = c("taxa.labels", "richness", "frequency",
  "sample.pool", "phylogeny.pool", "independentswap", "trialswap"),
  abundance.weighted = FALSE, runs = 999, iterations = 1000)
}
\arguments{
\item{samp}{Community data matrix}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{null.model}{Null model to use (see Details section for description)}

\item{abundance.weighted}{Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)}

\item{runs}{Number of randomizations}

\item{iterations}{Number of iterations to use for each randomization (for independent swap and trial null models)}
}
\value{
A list of results:
\itemize{
  \item ntaxa - Number of taxa in community
  \item comdist.obs - Observed mpd between community
  \item comdist.rand.mean - Mean mpd between null communities
  \item comdist.rand.sd - Standard deviation of mpd between null communities
  \item comdist.obs.rank - Rank of observed mpd vs. null mpd
  \item comdist.obs.z - Standardized effect size of mpd vs. null mpd (= (comdist.obs - comdist.rand.mean) / comdist.rand.sd, equivalent to -betaNRI)
  \item comdist.obs.p - P-value (quantile) of observed mpd vs. null communities (= comdist.obs.rank / runs + 1)
  \item runs - Number of randomizations
}
}
\description{
Standardized effect size of MPD (mean pairwise distance) separating taxa in two communities, a measure of phylogenetic beta diversity
}
\details{
Currently implemented null models (arguments to null.model):
\itemize{
 \item taxa.labels - Shuffle distance matrix labels (across all taxa included in distance matrix) 
 \item richness - Randomize community data matrix abundances within samples (maintains sample species richness) 
 \item frequency - Randomize community data matrix abundances within species (maintains species occurence frequency) 
 \item sample.pool - Randomize community data matrix by drawing species from pool of species occurring in at least one community (sample pool) with equal probability   
 \item phylogeny.pool- Randomize community data matrix by drawing species from pool of species occurring in the distance matrix (phylogeny pool) with equal probability   
 \item independentswap - Randomize community data matrix with the independent swap algorithm (Gotelli 2000) maintaining species occurrence frequency and sample species richness  
 \item trialswap - Randomize community data matrix with the trial-swap algorithm (Miklos & Podani 2004) maintaining species occurrence frequency and sample species richness  
}
}
\keyword{MPD}
\keyword{bMPD}
\keyword{bNRI}
\keyword{betaMPD}
\keyword{betaNRI}
\keyword{ses}
