% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.comdist2.R
\name{ses.comdist2}
\alias{ses.comdist2}
\title{Standardized effect size of inter-community MPD (betaMPD, betaNRI)}
\usage{
ses.comdist2(samp, dis, method = "quasiswap", fixedmar = "both",
  shuffle = "both", strata = NULL, mtype = "count", burnin = 0,
  thin = 1, abundance.weighted = FALSE, runs = 999)
}
\arguments{
\item{samp}{Community data matrix with samples as rows}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{method}{Character for method used for the swap algorithm ("swap", "tswap", "quasiswap", "backtrack"). If NULL no swap algorithm is applied (uses permatfull from vegan). If mtype="count" the "quasiswap", "swap", "swsh" and "abuswap" methods are available (see details).}

\item{fixedmar}{Character, stating which of the row/column sums should be preserved ("none", "rows", "columns", "both").}

\item{shuffle}{Character, indicating whether individuals ("ind"), samples ("samp") or both ("both") should be shuffled, see details.}

\item{strata}{Numeric vector or factor with length same as nrow(m) for grouping rows within strata for restricted permutations. Unique values or levels are used.}

\item{mtype}{Matrix data type, either "count" for count data, or "prab" for presence-absence type incidence data.}

\item{burnin}{Number of null communities discarded before proper analysis in sequential ("swap", "tswap") methods.}

\item{thin}{Number of discarded permuted matrices between two evaluations in sequential ("swap", "tswap") methods.}

\item{abundance.weighted}{Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)}

\item{runs}{Number of randomizations}
}
\value{
A list of results:
\itemize{
  \item ntaxa - Number of taxa in community
  \item comdist.obs - Observed mpd between communities
  \item comdist.rand.mean - Mean mpd between null communities
  \item comdist.rand.sd - Standard deviation of mpd between null communities
  \item comdist.obs.rank - Rank of observed mpd vs. null mpd
  \item comdist.obs.z - Standardized effect size of mpd vs. null mpd (= (comdist.obs - comdist.rand.mean) / comdist.rand.sd, equivalent to -betaNRI)
  \item comdist.obs.p - P-value (quantile) of observed mpd vs. null communities (= comdist.obs.rank / runs + 1)
  \item runs - Number of randomizations
}
}
\description{
Standardized effect size of MPD (mean pairwise distance) separating taxa in two communities, a measure of phylogenetic beta diversity
}
\details{
See permat (vegan) for detailed options
}
\keyword{MPD}
\keyword{bMPD}
\keyword{bNRI}
\keyword{betaMPD}
\keyword{betaNRI}
\keyword{ses}
