% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.comdistnt.R
\name{ses.comdistnt}
\alias{ses.comdistnt}
\title{Standardized effect size of inter-community MNTD (betaMNTD, betaNTI)}
\usage{
ses.comdistnt(samp, dis, null.model = c("taxa.labels", "richness",
  "frequency", "sample.pool", "phylogeny.pool", "independentswap", "trialswap"),
  abundance.weighted = FALSE, exclude.conspecifics = FALSE, runs = 999,
  iterations = 1000)
}
\arguments{
\item{samp}{Community data matrix with samples as rows}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{null.model}{Null model to use (see Details section for description)}

\item{abundance.weighted}{Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)}

\item{exclude.conspecifics}{Should conspecific taxa in different communities be exclude from MNTD calculations? (default = FALSE)}

\item{runs}{Number of randomizations}

\item{iterations}{Number of iterations to use for each randomization (for independent swap and trial null models)}
}
\value{
A list of results:
\itemize{
  \item ntaxa - Number of taxa in community
  \item comdistnt.obs - Observed mntd between communities
  \item comdistnt.rand.mean - Mean mntd between null communities
  \item comdistnt.rand.sd - Standard deviation of mntd between null communities
  \item comdistnt.obs.rank - Rank of observed mntd vs. null mntd
  \item comdistnt.obs.z - Standardized effect size of mntd vs. null mntd (= (comdistnt.obs - comdistnt.rand.mean) / comdistnt.rand.sd, equivalent to -betaNTI)
  \item comdistnt.obs.p - P-value (quantile) of observed mntd vs. null communities (= comdistnt.obs.rank / runs + 1)
  \item runs - Number of randomizations
}
}
\description{
Standardized effect size of MNTD (mean nearest taxon distance) separating taxa in two communities, a measure of phylogenetic beta diversity
}
\details{
Currently implemented null models (arguments to null.model):
\itemize{
 \item taxa.labels - Shuffle distance matrix labels (across all taxa included in distance matrix) 
 \item richness - Randomize community data matrix abundances within samples (maintains sample species richness) 
 \item frequency - Randomize community data matrix abundances within species (maintains species occurence frequency) 
 \item sample.pool - Randomize community data matrix by drawing species from pool of species occurring in at least one community (sample pool) with equal probability   
 \item phylogeny.pool- Randomize community data matrix by drawing species from pool of species occurring in the distance matrix (phylogeny pool) with equal probability   
 \item independentswap - Randomize community data matrix with the independent swap algorithm (Gotelli 2000) maintaining species occurrence frequency and sample species richness  
 \item trialswap - Randomize community data matrix with the trial-swap algorithm (Miklos & Podani 2004) maintaining species occurrence frequency and sample species richness  
}
}
\keyword{MNTD}
\keyword{bMNTD}
\keyword{bNTI}
\keyword{betaMNTD}
\keyword{betaNTI}
\keyword{ses}
