% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.comdistnt2.R
\name{ses.comdistnt2}
\alias{ses.comdistnt2}
\title{Standardized effect size of inter-community MNTD (betaMNTD, betaNTI)}
\usage{
ses.comdistnt2(samp, dis, method = "quasiswap", fixedmar = "both",
  shuffle = "both", strata = NULL, mtype = "count", burnin = 0,
  thin = 1, abundance.weighted = FALSE, exclude.conspecifics = FALSE,
  runs = 999)
}
\arguments{
\item{samp}{Community data matrix}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{method}{Character for method used for the swap algorithm ("swap", "tswap", "quasiswap", "backtrack"). If NULL no swap algorithm is applied (uses permatfull from vegan). If mtype="count" the "quasiswap", "swap", "swsh" and "abuswap" methods are available (see details).}

\item{fixedmar}{Character, stating which of the row/column sums should be preserved ("none", "rows", "columns", "both").}

\item{shuffle}{Character, indicating whether individuals ("ind"), samples ("samp") or both ("both") should be shuffled, see details.}

\item{strata}{Numeric vector or factor with length same as nrow(m) for grouping rows within strata for restricted permutations. Unique values or levels are used.}

\item{mtype}{Matrix data type, either "count" for count data, or "prab" for presence-absence type incidence data.}

\item{burnin}{Number of null communities discarded before proper analysis in sequential ("swap", "tswap") methods.}

\item{thin}{Number of discarded permuted matrices between two evaluations in sequential ("swap", "tswap") methods.}

\item{abundance.weighted}{Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)}

\item{runs}{Number of randomizations}
}
\value{
A list of results:
\itemize{
  \item ntaxa - Number of taxa in community
  \item comdistnt.obs - Observed mntd between community
  \item comdistnt.rand.mean - Mean mntd between null communities
  \item comdistnt.rand.sd - Standard deviation of mntd between null communities
  \item comdistnt.obs.rank - Rank of observed mntd vs. null mntd
  \item comdistnt.obs.z - Standardized effect size of mntd vs. null mntd (= (comdistnt.obs - comdistnt.rand.mean) / comdistnt.rand.sd, equivalent to -betaNTI)
  \item comdistnt.obs.p - P-value (quantile) of observed mntd vs. null communities (= comdistnt.obs.rank / runs + 1)
  \item runs - Number of randomizations
}
}
\description{
Standardized effect size of MNTD (mean nearest taxon distance) separating taxa in two communities, a measure of phylogenetic beta diversity
}
\details{
See permat (vegan) for detailed options
}
\keyword{MNTD}
\keyword{bMNTD}
\keyword{bNTI}
\keyword{betaMNTD}
\keyword{betaNTI}
\keyword{ses}
