% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.mntd.par.R
\name{ses.mntd.par}
\alias{ses.mntd.par}
\title{Standardized effect size of MNTD}
\usage{
ses.mntd.par(samp, dis, null.model = c("taxa.labels", "richness",
  "frequency", "sample.pool", "phylogeny.pool", "independentswap",
  "trialswap"), abundance.weighted = FALSE, runs = 999,
  iterations = 1000, cores = 1)
}
\arguments{
\item{samp}{Community data matrix with samples as rows}

\item{dis}{Distance matrix (generally a phylogenetic distance matrix)}

\item{null.model}{Null model to use (see Details section for description)}

\item{abundance.weighted}{Should mean nearest taxon distances for each species be weighted by species abundance? (default = FALSE)}

\item{runs}{Number of randomizations}

\item{iterations}{Number of iterations to use for each randomization (for independent swap and trial null models)}

\item{cores}{Number of cores to use for parallel computing}
}
\value{
A data frame of results for each community
\itemize{
\item{ntaxa}{Number of taxa in community}
\item{mntd.obs}{Observed mntd in community}
\item{mntd.rand.mean}{Mean mntd in null communities}
\item{mntd.rand.sd}{Standard deviation of mntd in null communities}
\item{mntd.obs.rank}{Rank of observed mntd vs. null communities}
\item{mntd.obs.z}{Standardized effect size of mntd vs. null communities (= (mntd.obs - mntd.rand.mean) / mntd.rand.sd, equivalent to -NRI)}
\item{mntd.obs.p}{P-value (quantile) of observed mntd vs. null communities (= mntd.obs.rank / runs + 1)}
\item{runs}{Number of randomizations}  
}
}
\description{
Parallel calculation of standardized effect size of mean nearest taxon distances in communities. When used with a phylogenetic distance matrix, equivalent to -1 times the Nearest Taxon Index (NTI).
}
\details{
Faster than ses.mntd from \code{picante} when there are many samples and taxa.

Currently implemented null models (arguments to null.model):
\itemize{
 \item taxa.labels - Shuffle distance matrix labels (across all taxa included in distance matrix) 
 \item richness - Randomize community data matrix abundances within samples (maintains sample species richness) 
 \item frequency - Randomize community data matrix abundances within species (maintains species occurence frequency) 
 \item sample.pool - Randomize community data matrix by drawing species from pool of species occurring in at least one community (sample pool) with equal probability   
 \item phylogeny.pool- Randomize community data matrix by drawing species from pool of species occurring in the distance matrix (phylogeny pool) with equal probability   
 \item independentswap - Randomize community data matrix with the independent swap algorithm (Gotelli 2000) maintaining species occurrence frequency and sample species richness  
 \item trialswap - Randomize community data matrix with the trial-swap algorithm (Miklos & Podani 2004) maintaining species occurrence frequency and sample species richness  
}
}
\keyword{MNTD}
\keyword{NTI}
\keyword{ses}
