% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.permtest.R
\name{ses.permtest}
\alias{ses.permtest}
\title{Permutation test of z-matrix}
\usage{
ses.permtest(zmat, sampleGroups, R = 10000)
}
\arguments{
\item{zmat}{Symmetric matrix with z-values}

\item{sampleGroups}{Vector with the grouping of the samples in the same order as in zmat}

\item{R}{Number of permutations}
}
\value{
A dataframe with p-values (pval), fdr corrected p-values (pval.adj), and average z-values (avg) for each group and combination of groups.
}
\description{
Permutation test of z-matrix from ses.comdist(2), ses.comdistnt and ses.UniFrac.
Test if means within or between groups is higher or lower than the overall mean of the matrix
}
