% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ses.UniFrac.R
\name{ses.UniFrac}
\alias{ses.UniFrac}
\title{Standardized effect size of unifrac}
\usage{
ses.UniFrac(physeq, method = "taxa.labels", fixedmar = "both",
  shuffle = "both", strata = NULL, mtype = "count", burnin = 0,
  thin = 1, weighted = TRUE, normalized = TRUE, runs = 99, cores = 1)
}
\arguments{
\item{physeq}{phyloseq-class, containing at minimum a phylogenetic tree and otu table}

\item{method}{"taxa.labels" shuffles labels in phylogenetic tree (Ignores fixedmar, shuffle, strata, mtype). If NULL then no swap algorithm is applied (i.e. uses permatfull from vegan). Else the method used for the swap algorithm ("swap", "tswap", "quasiswap", "backtrack"). If mtype="count" the "quasiswap", "swap", "swsh" and "abuswap" methods are available (see details).}

\item{fixedmar}{Character, stating which of the row/column sums should be preserved ("none", "rows", "columns", "both").}

\item{shuffle}{Character, indicating whether individuals ("ind"), samples ("samp") or both ("both") should be shuffled.}

\item{strata}{Numeric vector or factor for grouping samples within strata for restricted permutations. Unique values or levels are used.}

\item{mtype}{Matrix data type, either "count" for count data, or "prab" for presence-absence type incidence data.}

\item{burnin}{Number of null communities discarded before proper analysis in sequential ("swap", "tswap") methods.}

\item{thin}{Number of discarded permuted matrices between two evaluations in sequential ("swap", "tswap") methods.}

\item{weighted}{Should unifrac be weighted by species abundance? (default = TRUE)}

\item{normalized}{(Optional). Logical. Should the output be normalized such that values range from 0 to 1 independent of branch length values? Default is TRUE. Note that (unweighted) UniFrac is always normalized by total branch-length, and so this value is ignored when weighted == FALSE.}

\item{runs}{Number of randomizations}

\item{cores}{Number of cores to use for UniFrac of null communities. Default is 1}
}
\value{
A list of results:
\itemize{
  \item unifrac.obs - Observed unifrac between communities
  \item unifrac.rand.mean - Mean unifrac between null communities
  \item unifrac.rand.sd - Standard deviation of unifrac between null communities
  \item unifrac.obs.rank - Rank of observed unifrac vs. null unifrac
  \item unifrac.obs.z - Standardized effect size of unifrac vs. null unifrac (= (unifrac.obs - unifrac.rand.mean) / unifrac.rand.sd)
  \item unifrac.obs.p - P-value (quantile) of observed unifrac vs. null communities (= unifrac.obs.rank / runs + 1)
}
}
\description{
Standardized effect size of unifrac
}
\details{
See permat (vegan) for detailed options on permutation
}
\keyword{ses}
\keyword{unifrac}
