% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UniFrac.multi.R
\name{UniFrac.multi}
\alias{UniFrac.multi}
\title{Run \code{UniFrac} multiple times in parallel and take the average}
\usage{
UniFrac.multi(physeq, R = 100, seed = 42, cores = 1, ...)
}
\arguments{
\item{physeq}{Phyloseq object. Required}

\item{R}{Number of times to repeat calculation. Default 100}

\item{seed}{Random seed for reproducibility. Default 42}

\item{cores}{Number of cores to use for parallel computing. Default 1 aka not parallel}

\item{...}{Additional arguments passed to the \code{UniFrac} function}
}
\value{
A distance object with the average \code{UniFrac} distances
}
\description{
With unrooted phylogenies \code{UniFrac} sets the root randomly on the tree. The position of the root affects the results.
The function runs UniFrac multiple times, with different roots, and takes the average to smooth potential bias.
}
