% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WdS.R
\name{WdS.test}
\alias{WdS.test}
\title{Permutation test of Wd* - robust distance-based multivariate analysis of variance}
\usage{
WdS.test(dm, f, nrep = 999, strata = NULL)
}
\arguments{
\item{dm}{Distance matrix}

\item{f}{Factor}

\item{nrep}{Number of permutations}

\item{strata}{Factor for permuting in strata}
}
\value{
A list with a p-value, test statitic, and number of permutation
}
\description{
Robust distance-based multivariate analysis of variance (https://doi.org/10.1186/s40168-019-0659-9)
}
\details{
This code is taken from https://github.com/alekseyenko/WdStar/
}
\keyword{distance}
