% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adonis_OmegaSq.R
\name{adonis_OmegaSq}
\alias{adonis_OmegaSq}
\title{Calculate (partial) Omega-squared (effect-size calculation) for PERMANOVA and add it to the input object}
\usage{
adonis_OmegaSq(adonisOutput, partial = TRUE)
}
\arguments{
\item{adonisOutput}{An adonis object}

\item{partial}{Should partial omega-squared be calculated (sample size adjusted). Default TRUE}
}
\value{
Original adonis object with the (partial) Omega-squared values added
}
\description{
Calculate (partial) Omega-squared (effect-size calculation) for PERMANOVA and add it to the input object
}
