% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_venn.R
\name{ps_venn}
\alias{ps_venn}
\title{Make Venn diagram of shared taxa (ASVs, OTUs) across sample groups}
\usage{
ps_venn(
  ps,
  group,
  fraction = 0,
  weight = FALSE,
  relative = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{ps}{A phyloseq object}

\item{group}{The grouping factor. Should match variable in sample_data(ps)}

\item{fraction}{The fraction (0 to 1) of samples in a group in which the taxa should be present to be included in the count.}

\item{weight}{If TRUE, the overlaps are weighted by abundance}

\item{relative}{Should abundances be made relative}

\item{plot}{If TRUE return a plot, if FALSE return a list with shared and unique taxa}

\item{...}{Additional arguments}
}
\value{
An venn plot
}
\description{
Make Venn diagram of shared taxa (ASVs, OTUs) across sample groups from a phyloseq object. Overlap can be weighted by relative abundance
#' 
Any further arguments to this function are passed to the plot.venn function from the eulerr package. This can be used to change colors, fonts, and other graphical parameters.
For example: ps_venn(phy, "Time", quantities = list(type=c("percent","counts"), font = 2), labels = list(cex = 2), col = "red", fill = c("red","blue","green"))
}
\keyword{diagram}
\keyword{venn}
