% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_venn.R
\name{ps_venn}
\alias{ps_venn}
\title{Make Venn diagram of shared taxa (ASVs, OTUs) across sample groups}
\usage{
ps_venn(
  ps,
  group,
  fraction = 0,
  weight = FALSE,
  type = "percent",
  relative = TRUE,
  plot = TRUE,
  ...
)
}
\arguments{
\item{ps}{A phyloseq object}

\item{group}{The grouping factor. Should match variable in sample_data(ps)}

\item{fraction}{The fraction (0 to 1) of samples in a group in which the taxa should be present to be included in the count.}

\item{weight}{If TRUE, the overlaps are weighted by abundance}

\item{type}{"percent" or "counts"}

\item{relative}{Should abundances be made relative}

\item{plot}{If TRUE return a plot, if FALSE return a vector with number of shared and unique taxa}

\item{...}{Additional arguments}
}
\value{
An venn plot
}
\description{
Make Venn diagram of shared taxa (ASVs, OTUs) across sample groups from a phyloseq object. Overlap can be weighted by relative abundance
}
\keyword{diagram}
\keyword{venn}
